requirements.txt
pandas
numpy
scikit-learn
requests
PyYAML
vaderSentiment
joblib

